+++ etc/scripts/benchmark.sh
@@ -1,4 +1,5 @@
 #!/bin/bash
+set -eo pipefail
 if [ "$( echo "$1" | grep -q '[A-Za-z]' )" ]; then
   # first arg names the variable that JAVA_HOME is copied from
   export JAVA_HOME=${!1} # Bashism (https://github.com/koalaman/shellcheck/wiki/SC2039)
@@ -17,8 +18,6 @@ mvn -B -DskipTests -Darguments=-DskipTests\
 cd ../benchmark &&\
 mvn -B -DskipTests ${MAYBE_ANDROID_FLAG} package &&\
 cd target &&\
-"${JAVA_BIN}" ${JAVA_OPTS} -jar benchmarks.jar $2 -jvm "${JAVA_BIN}" -f 1 -t 1 -foe true -v EXTRA 2>&1 |\
-    tee benchmark_results_one_thread.txt &&\
-"${JAVA_BIN}" ${JAVA_OPTS} -jar benchmarks.jar $2 -jvm "${JAVA_BIN}" -f 1 -t 2 -foe true -v EXTRA 2>&1 |\
-    tee benchmark_results_two_threads.txt &&\
+"${JAVA_BIN}" ${JAVA_OPTS} -jar benchmarks.jar $2 -jvm "${JAVA_BIN}" -f 1 -t 1 -foe true -v EXTRA 2>&1
 cd ../..
+